% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{convert_data}
\alias{convert_data}
\title{convert data}
\usage{
convert_data(data, header)
}
\arguments{
\item{data}{data}

\item{header}{header}
}
\value{
tibble
}
\description{
convert data
}
\keyword{internal}
