% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{convert_temperature}
\alias{convert_temperature}
\title{convert temperature value}
\usage{
convert_temperature(v, span)
}
\arguments{
\item{v}{column of data which stores temperature}

\item{span}{returned value of parse_span()}
}
\value{
numeric
}
\description{
convert temperature value
}
\keyword{internal}
