% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{convert_variable}
\alias{convert_variable}
\title{convert each variables}
\usage{
convert_variable(ch, v, header)
}
\arguments{
\item{ch}{channel name}

\item{v}{variable; column of data}

\item{header}{header}
}
\value{
atomic vector
}
\description{
convert each variables
}
\keyword{internal}
