% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{convert_voltage}
\alias{convert_voltage}
\title{convert voltage value}
\usage{
convert_voltage(v, range)
}
\arguments{
\item{v}{column of data which stores voltage}

\item{range}{returned value of parse_range()}
}
\value{
numeric
}
\description{
convert voltage value
}
\keyword{internal}
