% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\docType{data}
\name{GBD_OF_V_L}
\alias{GBD_OF_V_L}
\title{value represent overflow}
\format{An object of class \code{numeric} of length 1.}
\usage{
GBD_OF_V_L
}
\description{
lower: -110% of lower full scale (-7FFF)
}
\keyword{internal}
