% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\docType{data}
\name{GBD_OF_V_U}
\alias{GBD_OF_V_U}
\title{Value represent overflow}
\format{An object of class \code{numeric} of length 1.}
\usage{
GBD_OF_V_U
}
\description{
Upper: 110% of upper full scale (+7FCC)
}
\keyword{datasets}
