% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{get_datetime_column}
\alias{get_datetime_column}
\title{get datetime column}
\usage{
get_datetime_column(header)
}
\arguments{
\item{header}{header}
}
\value{
POSIXct
}
\description{
get datetime column based on Stop Time (header$Measure$$Time > Stop) and Trigger Time ($Measure$$Time > Trigger). tz= "UTC"
}
\keyword{internal}
