% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{parse_amp}
\alias{parse_amp}
\title{parse Amp}
\usage{
parse_amp(Amp, CH)
}
\arguments{
\item{Amp}{header$Amp}

\item{CH}{Channel}
}
\value{
named character
}
\description{
offset: undocumented
}
\keyword{internal}
