% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{parse_header}
\alias{parse_header}
\title{parse header region}
\usage{
parse_header(header.str)
}
\arguments{
\item{header.str}{header reagion read by read_header()}
}
\value{
list
}
\description{
parse header region
}
