% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{parse_span_ad}
\alias{parse_span_ad}
\title{parse span ad}
\usage{
parse_span_ad(span, range)
}
\arguments{
\item{span}{header$Measure$Span[[CH]]}

\item{range}{returned value of parse_range()}
}
\value{
list
}
\description{
parse span ad
}
