% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gbd.R
\name{sampling_interval}
\alias{sampling_interval}
\title{get sampling interval}
\usage{
sampling_interval(s)
}
\arguments{
\item{s}{header$Common$$Data > Sample}
}
\value{
Period
}
\description{
get sampling interval
}
